Refactor: extract cipher_view class
